pyproject.toml
[project]
name = "demovietmindai"
version = "0.1.0"
description = "Agentic RAG with LangGraph for DASS21 and Knowledge Search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "google-genai>=1.55.0",
    "numexpr>=2.14.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-mongodb>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pymongo>=4.10.0",
    "motor>=3.6.0",
    "google-api-python-client>=2.149.0",
    "pydantic>=2.0.0",
]
